+++ src/components/level/toolbars/TilesToolbar/TilesToolbar.tsx
@@ -1,9 +1,8 @@
 import { FC, Fragment, useMemo } from "react";
 import cn from "classnames";
 import { useStore } from "effector-react";
-import { getDriver, getTilesForToolbar } from "drivers";
-import { $tileIndex, setTile } from "models/levels";
-import { $currentDriverName } from "models/levelsets";
+import { getTilesForToolbar } from "drivers";
+import { $drvTileRender, $drvTiles, $tileIndex, setTile } from "models/levels";
 import { Button, Toolbar } from "ui/button";
 import { ColorType, ContainerProps } from "ui/types";
 import cl from "./TilesToolbar.module.scss";
@@ -11,9 +10,16 @@ import cl from "./TilesToolbar.module.scss";
 interface Props extends ContainerProps {}
 
 export const TilesToolbar: FC<Props> = ({ className, ...rest }) => {
-  const driverName = useStore($currentDriverName)!;
-  const { tiles, TileRender } = getDriver(driverName)!;
-  const tilesSorted = useMemo(() => getTilesForToolbar(tiles), [tiles]);
+  const TileRender = useStore($drvTileRender)!;
+  const tiles = useStore($drvTiles)!;
+  const tilesSorted = useMemo(
+    () =>
+      getTilesForToolbar(tiles).filter(
+        ([, { value, metaTile }]) =>
+          !metaTile || metaTile.primaryValue === value,
+      ),
+    [tiles],
+  );
 
   const tileIndex = useStore($tileIndex);
   const handleTile = useMemo(
@@ -26,12 +32,12 @@ export const TilesToolbar: FC<Props> = ({ className, ...rest }) => {
 
   return (
     <Toolbar {...rest} className={cn(cl.root, className)}>
-      {tilesSorted.map(([i, { title, value }]) => (
-        <Fragment key={value ?? `?${i}`}>
+      {tilesSorted.map(([i, { title, value, metaTile }]) => (
+        <Fragment key={value}>
           {i > 0 && <wbr />}
           <Button
-            title={title}
-            icon={<TileRender tile={value} />}
+            title={metaTile?.title ?? title}
+            icon={metaTile?.icon ?? <TileRender tile={value} />}
             className={cn(cl.btn, i === tileIndex && cl._current)}
             asLink
             uiColor={ColorType.WARN}